+++ tox.ini
@@ -177,7 +177,7 @@ deps =
 
     ase
     vesin
-    nvalchemi-toolkit-ops
+    nvalchemi-toolkit-ops >=0.3.0,<0.4
 
     # for symmetrized calculator
     scipy
@@ -211,6 +211,7 @@ deps =
     vesin
     ase
     torch-sim-atomistic
+    nvalchemi-toolkit-ops >=0.3.0,<0.4
 
 changedir = python/metatomic_torchsim
 commands =
